pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0", "setuptools", "wheel"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ds-mono-repo"
version = "0.1.0"
description = ""
authors = ["Matt Elgazar <melgazar9@gmail.com>"]
readme = "README.md"
packages = [{include = "ds_core"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.2"
matplotlib = "^3.8.3"
numpy = "^1.26.4"
pandas = "^2.2.3"
plotly = "^5.19.0"
pyyaml = "^6.0.1"
seaborn = "^0.13.2"
scikit-learn = "^1.4.1.post1"
yagmail = "^0.15.293"
ds-core = {path = "ds_core"}
flatten-json = "^0.1.14"
dill = "^0.3.8"
feather-format = "^0.4.1"
catboost = "^1.2.2"
xgboost = "^2.0.3"
category-encoders = "^2.6.3"
imbalanced-learn = "^0.12.0"
feature-engine = "^1.6.2"
tqdm = "^4.66.2"
waitress = "^3.0.2"
apscheduler = "^3.11.0"
psycopg2 = "^2.9.10"
yfinance = "^0.2.55"
kagglehub = "^0.3.11"
sqlalchemy = "^2.0.40"
polars = "^1.29.0"
setuptools = "^80.4.0"
